+++ src/lib.rs
@@ -13,6 +13,7 @@ mod stage;
 pub mod basic_obj;
 pub mod fxaa;
 pub mod line;
+pub mod particle;
 pub mod pipeline;
 pub mod screen_quad;